share S3StorageService
