+++ SendgridParquet.Shared/SendgridParquet.Shared.csproj
@@ -5,4 +5,10 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="9.0.0" />
+    <PackageReference Include="ZLogger" Version="2.5.2" />
+  </ItemGroup>
+
 </Project>
