+++ SendgridParquetViewer/Components/_Imports.razor
@@ -13,5 +13,5 @@
 @using SendgridParquetViewer.Components.Account
 @using SendgridParquetViewer.Components.Layout
 @using SendgridParquetViewer.Services
-@using SendgridParquetViewer.Options
+@using SendgridParquet.Shared
 @using Microsoft.FluentUI.AspNetCore.Components
